Rename module file and update comments
The srt.yml module file was misplaced in the 'bad' directory and has been moved to its correct location. Additionally, a comment was added to the file to explain the requirement for a more recent srt-devel package in CentOS 7 while building nginx-modules. This change also addresses the 'srt_listen_callback' error during the EL7 build process.